Adding BaseKafkaDataWriterBuilder to simplify KafkaWriterBuilders
